{"schema":"decodex/commit/1","summary":"Split MCP planning server modules","authority":"XY-1115"}
